Refactor hybrid test to base class
