+++ test/connectors/test_gpu_hybrid_torch_models.py
@@ -1,14 +1,43 @@
-from qiskit_machine_learning import QiskitMachineLearningError
-from test import gpu
-from test.connectors.test_hybrid_torch_models import TestHybridTorchModels
-from test.connectors.test_torch_connector import TestTorchConnector
+# This code is part of Qiskit.
+#
+# (C) Copyright IBM 2022.
+#
+# This code is licensed under the Apache License, Version 2.0. You may
+# obtain a copy of this license in the LICENSE.txt file in the root directory
+# of this source tree or at http://www.apache.org/licenses/LICENSE-2.0.
+#
+# Any modifications or derivative works of this code must retain this
+# copyright notice, and modified files need to carry a notice indicating
+# that they have been altered from the originals.
+
+"""Test Torch GPU Torch Connector."""
+
+import unittest
+from test import QiskitMachineLearningTestCase, gpu
+from test.connectors.test_torch_models import TestTorchModels
+from qiskit.utils import optionals
+import qiskit_machine_learning.optionals as _optionals
 
 
-@gpu
-class TestGPUTorchConnector(TestHybridTorchModels):
+class TestGPUTorchConnector(QiskitMachineLearningTestCase, TestTorchModels):
+    """GPU Hybrid model tests for Torch Connector."""
+
+    @unittest.skipUnless(optionals.HAS_AER, "qiskit-aer is required to run this test")
+    @unittest.skipIf(not _optionals.HAS_TORCH, "PyTorch not available.")
+    @gpu
+    def setUp(self):
+        super().setup_test()
+        super().setUp()
+
     def _get_device(self):
         import torch
 
         if not torch.cuda.is_available():
-            raise QiskitMachineLearningError("CUDA is not available")
+            # raise test exception
+            self.skipTest("CUDA is not available")
+            return None
         return torch.device("cuda")
+
+
+if __name__ == "__main__":
+    unittest.main()
