+++ test/connectors/test_hybrid_torch_models.py
@@ -14,172 +14,25 @@
 
 import unittest
 from test import QiskitMachineLearningTestCase
-import numpy as np
-from ddt import ddt, idata
+from test.connectors.test_torch_models import TestTorchModels
 
-import qiskit
-from qiskit import QuantumCircuit
-from qiskit.circuit.library import RealAmplitudes, ZZFeatureMap
-from qiskit.utils import QuantumInstance, algorithm_globals, optionals
+from qiskit.utils import optionals
 import qiskit_machine_learning.optionals as _optionals
-from qiskit_machine_learning.neural_networks import CircuitQNN, TwoLayerQNN, NeuralNetwork
-from qiskit_machine_learning.connectors import TorchConnector
 
 
-@ddt
-class TestHybridTorchModels(QiskitMachineLearningTestCase):
+class TestHybridTorchModels(QiskitMachineLearningTestCase, TestTorchModels):
     """Hybrid model tests for Torch Connector."""
 
     @unittest.skipUnless(optionals.HAS_AER, "qiskit-aer is required to run this test")
     @unittest.skipIf(not _optionals.HAS_TORCH, "PyTorch not available.")
     def setUp(self):
+        super().setup_test()
         super().setUp()
-        algorithm_globals.random_seed = 12345
-        # specify quantum instances
-        self.sv_quantum_instance = QuantumInstance(
-            qiskit.Aer.get_backend("aer_simulator_statevector"),
-            seed_simulator=algorithm_globals.random_seed,
-            seed_transpiler=algorithm_globals.random_seed,
-        )
-        # pylint: disable=no-member
-        self.qasm_quantum_instance = QuantumInstance(
-            qiskit.providers.aer.AerSimulator(),
-            shots=100,
-            seed_simulator=algorithm_globals.random_seed,
-            seed_transpiler=algorithm_globals.random_seed,
-        )
-        import torch
-
-        torch.manual_seed(algorithm_globals.random_seed)
-
-    def _create_network(self, qnn: NeuralNetwork, output_size: int):
-        from torch import cat
-        from torch.nn import Linear, Module
-        import torch.nn.functional as F
-
-        # set up dummy hybrid PyTorch module
-        class Net(Module):
-            """PyTorch nn module."""
-
-            def __init__(self):
-                super().__init__()
-                self.fc1 = Linear(4, 2)
-                self.qnn = TorchConnector(
-                    qnn, np.array([0.1] * qnn.num_weights)
-                )  # Apply torch connector
-                self.fc2 = Linear(output_size, 1)  # shape depends on the type of the QNN
-
-            def forward(self, x):
-                """Forward pass."""
-                x = F.relu(self.fc1(x))
-                x = self.qnn(x)  # apply QNN
-                x = self.fc2(x)
-                return cat((x, 1 - x), -1)
-
-        return Net()
 
     def _get_device(self):
         import torch
 
         return torch.device("cpu")
-
-    def _create_circuit_qnn(self) -> CircuitQNN:
-        output_shape, interpret = 2, lambda x: f"{x:b}".count("1") % 2
-        num_inputs = 2
-
-        feature_map = ZZFeatureMap(num_inputs)
-        ansatz = RealAmplitudes(num_inputs, entanglement="linear", reps=1)
-
-        qc = QuantumCircuit(num_inputs)
-        qc.append(feature_map, range(num_inputs))
-        qc.append(ansatz, range(num_inputs))
-
-        qnn = CircuitQNN(
-            qc,
-            input_params=feature_map.parameters,
-            weight_params=ansatz.parameters,
-            input_gradients=True,  # for hybrid qnn
-            interpret=interpret,
-            output_shape=output_shape,
-            quantum_instance=self.sv_quantum_instance,
-        )
-        return qnn
-
-    def _create_opflow_qnn(self) -> TwoLayerQNN:
-        num_inputs = 2
-
-        # set up QNN
-        qnn = TwoLayerQNN(
-            num_qubits=num_inputs,
-            quantum_instance=self.sv_quantum_instance,
-            input_gradients=True,  # for hybrid qnn
-        )
-        return qnn
-
-    @idata(["opflow", "circuit_qnn"])
-    def test_hybrid_batch_gradients(self, qnn_type: str):
-        """Test gradient back-prop for batch input in a qnn."""
-        import torch
-        from torch.nn import MSELoss
-        from torch.optim import SGD
-
-        if qnn_type == "opflow":
-            qnn = self._create_opflow_qnn()
-            output_size = 1
-        elif qnn_type == "circuit_qnn":
-            qnn = self._create_circuit_qnn()
-            output_size = 2
-        else:
-            raise ValueError("Unsupported QNN type")
-
-        model = self._create_network(qnn, output_size=output_size)
-        device = self._get_device()
-        model.to(device)
-
-        # random data set
-        x = torch.rand((5, 4), device=device)
-        y = torch.rand((5, 2), device=device)
-
-        # define optimizer and loss
-        optimizer = SGD(model.parameters(), lr=0.1)
-        f_loss = MSELoss(reduction="sum")
-
-        # loss and gradients without batch
-        optimizer.zero_grad(set_to_none=True)
-        sum_of_individual_losses = 0.0
-        for x_i, y_i in zip(x, y):
-            output = model(x_i)
-            sum_of_individual_losses += f_loss(output, y_i)
-        sum_of_individual_losses.backward()
-        sum_of_individual_gradients = 0.0
-        for n, param in model.named_parameters():
-            # make sure gradient is not None
-            self.assertFalse(param.grad is None)
-            if n.endswith(".weight"):
-                sum_of_individual_gradients += np.sum(param.grad.numpy())
-
-        # loss and gradients with batch
-        optimizer.zero_grad(set_to_none=True)
-        output = model(x)
-        batch_loss = f_loss(output, y)
-        batch_loss.backward()
-        batch_gradients = 0.0
-        for n, param in model.named_parameters():
-            # make sure gradient is not None
-            self.assertFalse(param.grad is None)
-            if n.endswith(".weight"):
-                batch_gradients += np.sum(param.grad.numpy())
-
-        # making sure they are equivalent
-        self.assertAlmostEqual(
-            np.linalg.norm(sum_of_individual_gradients - batch_gradients), 0.0, places=4
-        )
-
-        self.assertAlmostEqual(
-            sum_of_individual_losses.detach().numpy(),
-            batch_loss.detach().numpy(),
-            places=4,
-        )
 
 
 if __name__ == "__main__":
